+++ packages/components-react/projects/react-ssr-wrapper/rollup.config.js
@@ -1,6 +1,7 @@
 import typescript from '@rollup/plugin-typescript';
 import generatePackageJson from 'rollup-plugin-generate-package-json';
 import resolve from '@rollup/plugin-node-resolve';
+import preserveDirectives from 'rollup-plugin-preserve-directives';
 
 const outputDir = '../../dist/react-wrapper/ssr';
 const input = 'src/public-api.ts';
@@ -10,6 +11,15 @@ const typescriptOpts = {
 };
 
 const external = ['@porsche-design-system/components-js', 'react', 'react/jsx-runtime'];
+
+// to silence warnings like
+// Module level directives cause errors when bundled, "use client" in "..." was ignored.
+// https://github.com/Ephem/rollup-plugin-preserve-directives#disabling-warnings
+const onwarn = (warning, warn) => {
+  if (warning.code !== 'MODULE_LEVEL_DIRECTIVE') {
+    warn(warning);
+  }
+};
 
 export default [
   {
@@ -21,6 +31,7 @@ export default [
       preserveModules: true,
     },
     plugins: [
+      preserveDirectives.default(),
       resolve(),
       typescript({
         ...typescriptOpts,
@@ -37,6 +48,7 @@ export default [
         },
       }),
     ],
+    onwarn,
   },
   {
     input,
@@ -46,6 +58,7 @@ export default [
       format: 'esm',
       preserveModules: true,
     },
-    plugins: [resolve(), typescript(typescriptOpts)],
+    plugins: [preserveDirectives.default(), resolve(), typescript(typescriptOpts)],
+    onwarn,
   },
 ];