+++ packages/components/CHANGELOG.md
@@ -16,8 +16,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 #### Added
 
+- React: `'use client';` directive is applied on all components for main and `ssr` sub-package
+  ([#2654](https://github.com/porsche-design-system/porsche-design-system/pull/2654))
 - Breakpoint customizable property `column` to `Segmented Control` to control the width of items
   ([#2652](https://github.com/porsche-design-system/porsche-design-system/pull/2652))
+
+#### Fixed
+
+- Fixes regression in `observeChildren` that affected nested components (e.g. incorrect rendering of nested `Tabs`).
+  ([#2649](https://github.com/porsche-design-system/porsche-design-system/pull/2649))
 
 ### [3.3.0] - 2023-07-07